+++ tests/Facades/MarkdownTest.php
@@ -14,7 +14,7 @@ namespace GrahamCampbell\Tests\Markdown\Facades;
 use GrahamCampbell\Markdown\Facades\Markdown;
 use GrahamCampbell\TestBenchCore\FacadeTrait;
 use GrahamCampbell\Tests\Markdown\AbstractTestCase;
-use League\CommonMark\CommonMarkConverter;
+use League\CommonMark\Converter;
 
 /**
  * This is the markdown facade test class.
@@ -52,7 +52,7 @@ class MarkdownTest extends AbstractTestCase
      */
     protected function getFacadeRoot()
     {
-        return CommonMarkConverter::class;
+        return Converter::class;
     }
 
     public function testConvertToHtml()